Raise exception if given an empty buffer when opening workbook
Prior to this change, when passing the name of an empty file to
Workbook.open, an exception would be raised because the underlying zip
library [does so](https://github.com/rubyzip/rubyzip/blob/e04c9cdbd8ddcca239c4628380dee7dd36de873b/lib/zip/file.rb#L100).

However, if you pass an empty buffer that can be read, prior to this
change, the workbook will attempt to read it and fail. It will fail
because when the initializer is attempting to set the styles, sheet IDs,
and relationships, it will not find the expected .xml file in the zip,
and therefore attempt to call `#get_input_stream` on `nil`, which will
fail.

Before determining those values, this modifies the initializer to check
to see when we get a buffer, which will pass the zip library's
checking of it being empty, if the `Zip::File` we receive has any size.

This makes the behavior consistent across different scenarios, where an
exception is raised. However, the *type* of exception differs. In the
buffer case, this will raise an exception from this library, as opposed
to the `Zip::Error` exception raised by that library itself.