Optimize for YJIT by using while loops
The code introduced in this change is not idiomatic ruby but improves performance in Ruby 3.3-rc1 and without YJIT enabled.

With YJIT the performance gain is about 5%, without YJIT it's 2%. YJIT is generally 15% faster in the basic Xsv benchmark.

For more information, see the code optimization tips in the [YJIT docs](https://github.com/ruby/ruby/blob/master/doc/yjit/yjit.md#code-optimization-tips)